Clean up code and fix documentation.
